requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pymongo==4.6.0
pydantic==2.4.2
python-dotenv==1.0.0
motor==3.3.1 